requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-crontab==0.7.1
djangorestframework==3.12.4
idna==3.2
mysqlclient==2.0.3
Pillow==8.3.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
schedule==1.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
